Show Heat links in a panel group on project dashboard
The group title is 'Orchestration' and currently the only panel
is 'Stacks'.

In the previous Thermal demo there was a separate dashboard called
'Heat'. I think a panel group is more appropriate, since Heat stacks
exist in the context of a project, just like Compute and Object Store.

Also, users will currently see 'Orchestration' as the title of this
feature, since Heat is project jargon.