Implementation of a Heat stacks UI.
This change implements a UI to launch, manage and delete Heat stacks.

The launch screens are implemented with a workflow where the second
step is dynamically built from the template uploaded in the first step.

A significant portion of this change was derived from
the work Dan Radez <dradez@redhat.com> did on thermal:
https://github.com/steveb/heat-horizon

Implementes blueprint: heat-ui

Change-Id: I294e93bed6da9dd3553e8b4a6a1c09b7c165a555